% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.oc.combBB.R
\name{get.oc.combBB}
\alias{get.oc.combBB}
\title{Generate operating characteristics for drug combination trials}
\usage{
get.oc.combBB(BLRMspecs, target, p.true, ncohort, cohortsize, preferred.doses, n.earlystop=NULL, startdose=c(1, 1),
                   titration=FALSE,p.saf=0.6*target, p.tox=1.4*target, cutoff.eli=0.95,
                   extrasafe=FALSE,offset=0.05, ntrial=1000, mtd.contour=FALSE,
                   boundMTD=FALSE, seed=6)
}
\arguments{
\item{BLRMspecs}{The priors and parameters needed for BLRM (see package blrm for detailed specification)}

\item{target}{the target DLT rate}

\item{p.true}{a \code{J*K} matrix \code{(J<=K)} containing the true toxicity probabilities of
combinations with \code{J} dose levels of agent A and \code{K} dose levels of agent B}

\item{ncohort}{a \code{1*J} vector specifying the number of cohorts for each of \code{J} subtrials
if \code{mtd.contour=TRUE}; Otherwise, a scalar specifying the total number of cohorts for
the trial.}

\item{cohortsize}{the cohort size}

\item{preferred.doses}{the matrix that specifies the admissible dose combinations:
\code{1} for included dose combinations and \code{-1} for excluded
dose combinations. The starting dose must be coded as \code{1}.}

\item{n.earlystop}{the early stopping parameter. If the number of patients treated at the current
dose reaches \code{n.earlystop}, stop the trial or subtrial and select the MTD based on
the observed data. When the waterfall design is used to find the MTD contour,
\code{n.earlystop=12} by default.}

\item{startdose}{the starting dose combination level for drug combination trial}

\item{titration}{set \code{titration=TRUE} to perform dose escalation with cohort size = 1 to accelerate dose escalation at the begining of the trial.}

\item{p.saf}{the highest toxicity probability that is deemed subtherapeutic (i.e. below the MTD)
such that dose escalation should be undertaken.
The default value is \code{p.saf=0.6*target}.}

\item{p.tox}{the lowest toxicity probability that is deemed overly toxic such that deescalation
is required. The default value is \code{p.tox=1.4*target}.}

\item{cutoff.eli}{the cutoff to eliminate an overly toxic dose for safety. We recommend the
default value of (\code{cutoff.eli=0.95}) for general use.}

\item{extrasafe}{set \code{extrasafe=TRUE} to impose a more stringent stopping rule}

\item{offset}{a small positive number (between 0 and 0.5) to control how strict the stopping
rule is when \code{extrasafe=TRUE}. A larger value leads to a more strict stopping
rule. The default value \code{offset=0.05} generally works well.}

\item{ntrial}{the total number of trials to be simulated}

\item{mtd.contour}{set \code{mtd.contour=TRUE} to select the MTD contour (claiming multiple MTDs).
Otherwise, BOIN design is used to search for a single MTD.}

\item{boundMTD}{set \code{boundMTD=TRUE} to impose the condition: the isotonic estimate of toxicity
probability for the selected MTD must be less than de-escalation boundary.}

\item{explore_deescalation}{TRUE or FALSE. Whether to go to unexplored dose in de-escalation, disregarding the posterior probability.}

\item{seed}{the random seed for simulation}
}
\value{
\code{get.oc.combP()} returns the operating characteristics of the BOIN combination or
waterfall design as a list. For the BOIN combination design, including:
(1) true toxicity probability at each dose level (\code{$p.true}),
(2) admissibility indicator at each dose level (\code{$preferred.doses}),
(3) selection percentage at each dose level (\code{$selpercent}),
(4) the number of patients treated at each dose level (\code{$npatients})
(5) the number of toxicities observed at each dose level (\code{$ntox}),
(6) the total number of patients in the trial (\code{$totaln}),
(7) the total number of toxicities observed for the trial (\code{$totaltox})
(8) the percentage of correct selection (\code{$pcs}),
(9) the percentage of acceptable selection (\code{$pas}),
(10) the total percentage of patients treated at the MTD (\code{$npercent}).
(11) the percentage of early stopping without selecting the MTD (\code{$percentstop})
}
\description{
Obtain the operating characteristics of the BOIN-CombP design for drug combination
 trials. The BOIN-CombP design is to find a MTD within specified drug combinations in the dose matrix.
}
\details{
The operating characteristics of the BOIN-CombP design are generated by
simulating trials under the prespecified true toxicity probabilities of the investigational dose
combinations. If \code{titration=TRUE}, we perform dose escalation with cohort size = 1 at the begining of the trial:
starting from \code{startdose}, if no toxicity is observed, we escalate the dose;
otherwise, the titration is completed and we switch to cohort size = \code{cohortsize}.
Titration accelerates the dose escalation and is useful when low doses are believed to be safe.

The dose combinations to be considered are specified by \code{preferred.doses},
where dose combinations coded as \code{1} are included and dose combinations
coded as \code{-1} are excluded from escalation and de-escalation. Among the
admissible neighboring dose combinations, the next dose is chosen based on the
posterior probability of being in the optimal interval.


The BOIN-CombP design has two built-in stopping rules:
(1) stop the trial/subtrial if the lowest dose is eliminated due to toxicity, and no dose should
be selected as the MTD; and (2) stop the trial/subtrial and select the MTD if the number of
patients treated at the current dose reaches \code{n.earlystop}. The first stopping rule is a safety
rule to protect patients from the case in which all doses are overly toxic. The rationale for
the second stopping rule is that when there is a large number (i.e., \code{n.earlystop}) of
patients assigned to a dose, it means that the dose-finding algorithm has approximately converged.
Thus, we can stop the trial/subtrial early and select the MTD to save sample size and reduce the
trial duration.


 For some applications, investigators may prefer a more strict safety stopping rule than rule
 (1) for extra safety when the lowest dose is overly toxic.
 This can be achieved by setting \code{extrasafe=TRUE},
 which imposes the following more strict safety stopping rule:
 stop the trial if (i) the number of patients treated at the lowest dose \code{>=3},
and (ii) \eqn{Pr(toxicity\ rate\ of\ the\ lowest\ dose > \code{target} | data) > \code{cutoff.eli}-\code{offset}}.
As a tradeoff, the strong stopping rule will decrease the MTD selection percentage
 when the lowest dose actually is the MTD.
}
\note{
We should avoid setting the values of \code{p.saf} and \code{p.tox} very close to the
     \code{target}. This is because the small sample sizes of typical phase I trials prevent us from
      differentiating the target DLT rate from the rates close to it. The default values provided by
      \code{get.oc()} are strongly recommended, and generally yield excellent operating characteristics.
}
\examples{

###### drug-combination trial ######

##### combination trial to find a single MTD  ######

S1.p.true <- matrix(c(0.02, 0.07, 0.10,
0.07, 0.10, 0.15,
0.10, 0.15, 0.30,
0.15, 0.30, 0.45,
0.30, 0.45, 0.55), nrow = 3, byrow = FALSE)
doses.cp <- matrix(c(1, 1, 1, 1, 1,
                    1, 1,  1, 1, 1,
                    1, 1, 1, 1, 1),
                  ncol = 5, byrow = TRUE)

## combo version
# prior
## drug 1
mean1 <- c(-1.0989, -0.1674)
se1 <- c(1.2770, 0.5713)
corr1 <- 0.5224
prior1 <- list(mean=mean1, se=se1, corr=corr1)
## drug 2
mean2 <- c(-2.9444, 0)
se2 <- c(2, 1)
corr2 <- 0
prior2 <- list(mean=mean2, se=se2, corr=corr2)
##'interaction between 2 drugs
prior3 <- list(mean=0, se=1.121)
##'combine three sets of priors
prior <- list(prior1, prior2, prior3)

#'parameters
seeds <- 1:2
nsamples <- 10000
burn_in <- 0.5

##'drug 1
drug1_name <- "DRUG-X"
dose1_unit <- "mg"
ref_dose1 <- 5
prov_dose1 <- c(1, 2.5, 5)

##'drug 2
drug2_name <- "DRUG-Y"
dose2_unit <- "mg"
ref_dose2 <- 400
prov_dose2 <- c(200, 250, 300, 350, 400)

category_bound <- c(0.16, 0.33)
category_name <- c("under-dosing", "targeted-toxicity", "over-dosing")
ewoc <- 0.30

#'combine to a list
BLRMspecs <- list(seeds=seeds, nsamples=nsamples, burn_in=burn_in,
                 drug1_name=drug1_name, dose1_unit=dose1_unit,
                 ref_dose1=ref_dose1, prov_dose1=prov_dose1,
                 drug2_name=drug2_name, dose2_unit=dose2_unit,
                 ref_dose2=ref_dose2, prov_dose2=prov_dose2,
                 #'dose1=dose1, dose2=dose2, n_pat=n_pat, dlt=dlt,
                 category_bound=category_bound, category_name=category_name,
                 ewoc=ewoc,
                 prior=prior)


oc.comb <- get.oc.combBB(BLRMspecs, target=0.3, p.true=S1.p.true, ncohort=10, cohortsize=3,
                        preferred.doses = doses.cp, n.earlystop=9, startdose=c(1,1), ntrial=1000)

summary(oc.comb)

}
\references{
Liu S. and Yuan, Y. (2015). Bayesian Optimal Interval Designs for Phase I Clinical
            Trials, \emph{Journal of the Royal Statistical Society: Series C}, 64, 507-523.

           Lin R. and Yin, G. (2017). Bayesian Optimal Interval Designs for Dose Finding in
           Drug-combination Trials, \emph{Statistical Methods in Medical Research}, 26, 2155-2167.

           Yan, F., Zhang, L., Zhou, Y., Pan, H., Liu, S. and Yuan, Y. (2020).BOIN: An R Package
           for Designing Single-Agent and Drug-Combination Dose-Finding Trials Using Bayesian Optimal
           Interval Designs. \emph{Journal of Statistical Software}, 94(13),1-32.<doi:10.18637/jss.v094.i13>.


           Zhang L. and Yuan, Y. (2016). A Simple Bayesian Design to Identify the Maximum
           Tolerated Dose Contour for Drug Combination Trials, \emph{Statistics in Medicine}, 35, 4924-4936.
}
\author{
Yuxuan Chen, Haiming Zhou, and Philip He.
}
