% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.boinP.R
\name{summary.boinP}
\alias{summary.boinP}
\title{Generate descriptive summary for objects returned by other functions.}
\usage{
\method{summary}{boinP}(object, ...)
}
\arguments{
\item{object}{the object returned by other functions.}

\item{...}{ignored arguments}
}
\value{
\code{summary()} prints the objects returned by other functions.
}
\description{
Generate descriptive summary for objects returned by other functions.
}
\details{
\code{summary()} prints the objects returned by other functions.
}
\examples{
## get the operating characteristics for BOIN design
S1.p.true <- matrix(c(0.02, 0.07, 0.10,
0.07, 0.10, 0.15,
0.10, 0.15, 0.30,
0.15, 0.30, 0.45,
0.30, 0.45, 0.55), nrow = 3, byrow = FALSE)
doses.cp <- matrix(c(1, -1, -1, -1, -1,
      1, 1, 1, 1, -1,
      1, 1, 1, 1, 1),
      ncol = 5, byrow = TRUE)
oc.comb <- get.oc.combP(target=0.3, p.true=S1.p.true, ncohort=10, cohortsize=3,
                               preferred.doses = doses.cp, n.earlystop=9, startdose=c(1,1), ntrial=1000)

summary(oc.comb)


}
\author{
Suyu Liu, Liangcai Zhang and Ying Yuan
}
